Added soil storage continuation between time steps
Currently there is no soil module in CHM. As a result, soil storage is
handled completely by the Infil_All module by increasing d.soil_storage
by d.GA_temp->final_storage in the GreenAmpt routine. The crack/Gray
module does not use soil_storage in its calculation.

When a soil module is added, this will have to be changed so that th
efinal storage is sent to the soil module and taken as an input.