Changed model to have parameters as public members
Each instance of this class will exist on every triangle. Parameters
which change with each triangle are stored as members of the class.
Currently, even parameters that are uniform are also stored this way.
(e.g., heat capacity, cp). Might need a new approach if these get
plentiful in the future.