Added initializers to PenmanMontieth model
Cp, Kappa, Pore size, and air-entry tension are all quantities accessed
from other parts of the code and are member variables of the
PenmanMontieth class.

These variables have been changed from double to const double& so that
they remain references to the existing values elsewhere in the code.

References are OK as long as we know that the instance of this class
is initialized after the original object is entered into memory and
destroyed before the original object.

In this case, this is quite safe because these objects are const members
of namespaces and never initialized in the normal sense.