PM_vars in PenmanMonteith model set as references
To ensure minimizing memory overhead, PM_vars are all set as references.

However, require a local copy of saturation vapour pressure (SVP) because it
is determined from a function from the Atmosphere namespace.

The reference has to be to something that exists and not the output of a
function (obviously).

SVP is a function of t, therefore a local copy of t is created so that
we don't have to dereference the face pointer twice. One in the
calculation of SVP and again in the initialization of PM_vars.

This slightly increases the memory overhead compared with dereferencing
face twice, once to calculate SVP and once to initialize PM_vars.

But the memory overhead of creating local copies on each triangle for
each member of PM_vars is definitely larger than the overhead of only
storing local copies of t and SVP.