Changed the output of veg_attribute to a reference
To be clear, the function still outputs double, not double&.

But instead the output variable is initialized as a reference to a
dummy variable, dummyvar, set to 0.0. dummyvar is a private member of
the face class.

Then if the output variable, result, is set to any other value, it will
be a reference to that value. All possible outputs are held in memory
for the duration of the simulation.