Changes and fixes in ET module
Typos and other errors fixed to get CHM to compile

Changes to the PenmanMonteith required changes to the ET module. All
inputs are references and initialized with the instance.

LAI, LAImax and vegetation_height now are stored as local copies because
of how face->veg_attribute is coded. Maybe changed later.