Modified soil_two_layer to interact with a DTO
DTO currently has every member as its own copy. This is memory
inefficient in two ways:

1. Output variables are held in memory even when the module is not run.

2. Domain wide variables are not references to a single instance in the
   module (one per process rather than one per triangle)

Reason 1 could be easily fixed by creating a sub struct to the DTO which
is accessed through a pointer. That pointer is deleted at the end of
each timestep.

Reason 2 is not a problem for this module because there aren't any
domain wide parameters. But it could be an issue for other modules.