Renamed the soil module files and change in func
soil.cpp is too close to physics/Soil.cpp, so renamed

func -> functionality.

soil_two_layer has a virtual function on its DTO to access if the face
is a lake or other body of water. In this commit I changed how this
worked. In the last change, I just copied the is_water function into the
soil_module which is awful.

What I do is define a struct in the data subclass that has a single
member. This single member is a reference of type soil_module (yes a
reference to the parent class, the module itself).

The data class has a member pointer that points to this struct. It is
left null in the data subclass definition.

In the init routine, this pointer is allocated as follows:

d.local_module = new soil_module::data::my_module(*this);

So that now the instance of the data class stored on the face object has
a pointer to its own parent.

I think this approach is terrible practice but it allows me to access
module_base functions while retaining the DTO interface between the
module itself and the various submodules.

A later version might have this function on a separate object that
could be accessed on the data class d.

Still must later actually get this to compile, problem for tomorrow...