Added soil functions to triangulation
New functions to triangulation for accessing soil related parameters.
The soil_attribute function, it is idenitical to veg_attribute except
its return type is a template and must be supplied as it is called as it
cannot be inferred by the underying function.

In veg_attribute, the output is always a double. But for soils the
output can be an interger, double or std::string.