Overloaded soil_attribute
The comment next to this function (not the definition but the
declaration) explains. But essentially, the fact that the Soils name
space in physics/Soils.h requires a string means I require a look-up
table. However, there are too many options to use a similar format to
"land_cover" that veg_attribute uses. there is probably a better way but
this quick and dirty method will fix things.