Modified LAI in PM ET submodule
When I originally ported this module, I made the choice to avoid
recalculating LAI which it exists already in CHM for simple canopy.
However, to ensure a match with CRHM I've modified this part of the
Penman-Monteith class where the LAI is computed based on a growth
factor, vegetation height, and maximum and minimum LAI values.

Luckily for this run the growth factor is equal to 1, to avoid
additional changes, I set this as 1 instead of adding it as a new input.