Adjusted soil and ET modules depends
I have finally properly coodinated the depends/provides with the modules
themselves. I had originally designed these depends/provides with other
modules in CHM in mind, but for now to match the CRHM inputs that I will
be using, I've adjusted them so that they match what the rest of the
modules expects.