Infil_All bug fix
The counter of major melts, `major_melt_count`, was being incremented
after the initialization even if the melt is not major.

Added a new function that checks whether the melt is major before
incrementing.

This is a function between the counter is incremented in two places and
this ensures it is normalized.