Fixes to Infil_All
Major melt counting fixed and now counts past the infDays threshold to
be consistent with CHM. major_melt_count is also an output, useful for
diagnostics.

Temperature for ice-lens calculation adjusted from the temperature
during the time step to the daily maximum temperature. An ice lens in
this way is only possible if the warmest temperature of the day is below
the threshold.