Changed use of snowmelt after crack model ends
The crack model, enabled by the frozen variable in the data subclass,
deals with snowmelt when SWE is large enough (usually the winter).
However, there is often snowmelt after this period ends. In CRHM, this
is set as infiltrated. The previous commit had it as runoff but this was
a misreading of the CRHM code. This change is correct.