Order of Operations bug of initial conditions fix
Initial soil conditions (e.g.; `soil_storage`) are set in the init
function in the soil_module::init. However, they were previously only
stored locally to the module, rather than also on the face. This led to
a problem where if a module depends on one of these values but runs
first, then it will try to receive this variable from face prior to it
being set by the soil_module::run function.

This fix ensures that the ET module is able to properly access soil
initial conditions at the first time step without needing to load it
from file itself.