Adjusted ET functions to have more sensible units
Original equations for the latent heat (lambda) of vapourization and the
psychrometric constant are reworked to have better units (Joules rather
than kilijoules or megajoules) and more true to their original form.

Such as the psychrometric constant, it depends on the heat capacity of
air but the original equation does not.

Arguments of these functions have been adjusted to be const as well.