Modified CHM code for unit tests
Unit tests compare CHM output of Evapotranspiration_All, soil_module and
Infil_All to CRHM output. Inputs to these modules are supplied by
point_mode after these adjustments.

core.cpp changes are done so that the code is able to properly find the
triangle. This is a unit test where it is based on a by-hand MESHER
output. CHM requires that you pinpoint a specific point on the MESH
domain and so if your triangle is 'wrong' it will fail.

CRHM runs K_Estimate *after* the SoilX module. Therefore, here, we move
The K_Estimate calculation in soil_two_layer to after the soil moisture
movement.

Note that it is still not after soil_ET and should be considered in
these tests.