XG Algorithm Added to compute Freeze/thaw fronts
New class XG_algorithm which exists to compute Freeze/Thaw fronts in the
soil.

For now, it is included directly in soil_module rather than its own
module because it has some level of coupling to the soil_module.

For example, it depends on this division of the soil into a recharge
layer and a lower layer. If this is later eliminated, then a separate
model could be written.

XG_algorithm is currently derived from a baseclass I_freeze_thaw_depths
but it does not make use of the interface in a meaningful way. It has a
virtual run function that it uses but a pointer to this baseclass is
never used and instead the XG_algorthm object is created directly. This
design is based on old ideas that could be tossed in the future but for
now this is what we have.

XG_algorithm has two classes defined within it, State and Param. State
holds the state of XG between time steps. Param holds the parameters.
Inputs to XG_algorithm are passed in the constructor (it is constructed
every time step, again, a design that should change later).

State also contains several functions which modify the state. These
functions represent the steps that the original XG algorithm does not
include (original paper below). For example, the computation of thermal
conductivities.

GoogleTests of State and XG_algorithm found in test_XG.cpp.

Changwei, Xie, and William A. Gough. “A Simple Thaw-Freeze Algorithm for a Multi-Layered Soil Using the Stefan Equation.” Permafrost and Periglacial Processes, vol. 24, no. 3, 2013, pp. 252–60. Wiley Online Library, https://doi.org/10.1002/ppp.1770.