Added XG algorithm and refactored soil_two_layer
XG_algorithm, port of CRHM module ClassXG included, as well as extensive
unit tests of the class and its components.

soil_two_layer refactored for more clear SRP. This was to enable more
rigorous testing with more encapsulation and separation of
responsibilities.

As googletests were updated, now includes googlemock functionality in
the CMakeLists, and used in test_soil which tests the objects in
soil_classes used by soil_two_layer