Optimizing base_step and deriving classes
base_step and derived classes previously had a single instance per face,
defined on the data class as a unique_ptr.

There are several drawbacks to this. First, unique_ptrs are on the heap
and typically slower than the stack. Each object requires an address of
8-16 bytes, plus an address for the template reference to the data
object. As shown in other branches (e.g., new_module/t_surface in the
file surface_temperature.cpp), due to the data template needed by the
sumodule data is an incomplete type on its declaration (since each is
declared as a member of the data class), we require a separate class to
actually hold the functions. This just further increases the complexity
of writing a module.

As shown by the changes presented here, we now have a single instance of
the submodule per module. In this case

```cpp
net_radiation net_rad;
```

and pass the data object in the execute function:

```cpp
void net_radiation::run(mesh_elem& face)
{
    // get data

    net_rad.execute(d);

    // set outputs
};
```

This radiaically simplifies the module itself, while also possibly
resulting in memory savings, performance improvements at the cost of
only a slight more verbose submodule function calls along with the very
slight overhead of the passing of an argument to the function.