Modify base_step for single-instance implementation
As designed, derived classes of base_step must have a separate instance
per triangle face. This is because it takes the data object as a
reference and there is a single data object per face. At least in the
current standard of implementation.

This new version takes a reference to the data object in the execute
function which means that a module only needs a single instance. This
has several savings.

1. Previously, we build the object in init as a `unique_ptr` using
   `make_unique`. This was done because it needed to be connected to the
data object at runtime.

2. Now we save memory for each pointer and for each reference to the data object.

3. No more pointer indirection for execute calls. Slight efficiency
   improvement.

4. Code is much more readable.

Cons:

1. Class is now functionally stateless. Parameters which are constant
   across faces and set at run-time can still be held on the object.
This means that in some cases, data object construction is more complex.
In the previous iteration, state variables that are not needed to be
known externally to the base_step derived class could have been stored
as a private member on the object directly. This is mostly OK because
the approach here would be to create public members of the data object
that are not functions and just raw variables.

Possibly others. I believe pros outweigh the cons