Added net_all_bad_lake to Evapotranspiration_All
net_all_bad_lake is only used by the evapotranspiration modules and
therefore is just included in it.

net_all_bad_lake is derived from base_step and therefore makes use of
data_base for caching. The data class is therefore also updated for this
purpose. Added Cache class for use by the data class parent class.

The rest of Evapotranspiration_All still uses older paradigms and will
be updated in the future.