New CHM module soil_water_unit_conversion
Computes:

1. volumetric water content

2. degree of saturation "saturation"

Volumetric water content is defined as the volume of water divided by
the volume of the soil, where the volume of the soil is the sum of the
volume of water, air and minerals. In order words, it is the fraction of
the soil that is water.

The module soil_module computes soil moisture content, soil_storage.
However, it only some of the moisture. In particular, it includes
moisture above the wilt point so in order to get the volumetric water
content and degree of saturation this must be accounted for.

The submodule soil_moisture_converter accounts for this. Setting
d.fractional_cutoff() > 0.0 achieves this.