Modified Infil_All to work with soil_water_unit_converter
soil_water_unit_converter outputs degree of saturation as
`soil_saturation`. `Infil_All` requires this value to be computed when
the soil first freezes and is used to define the infiltration regime.

`soil_storage_at_freeze` is now an `std::optional<double>` stored on the
data class. This way it can use the null checking aspects of pointers
without being one.