Modifying data_base derived class functions
Two main changes. `data` classes derived from `data_base` are expected
to implement functions to provide inputs to the submodules. I've removed
the const modifier on these functions.

The second change: implementations of `update_field` and `set_output`
must provide a lambda for cache access. This is done so that
init_cache() can be called within update_field and to avoid access cache
members prior to the cache being instantiated.