Changes to data_base usage and internal
Main changes:

1. More specific concepts. ValueRules is the rules required for the
   first argument to `update_field`. This concept ensures that this
input is callable and returns a lvalue reference. OutputRules is the
rules required for the first argument of the `set_output` function. It
first requires that it satisifies ValueRules and returns the same type
as the second argument.

2. Requiring that `set_ouput` and `update_field` first arguments must be
   callable and should be lambdas. This is done via the ValueRules and
OutputRules concepts. The reason for this is that previously the
functinos were written like:

```cpp
update_field(cache_->value,[this]() {return (*face)["value"_s]; });
```

This is problematic because this function is called before the cache_ is
allocated and therefore can lead to unexpected results. An option was to
move the allocation step (and check) outside of update_field but that
would require users to manually call it when writing `data` functions.
This creates extra boiler plate. Instead, the first argument must be a
lambda:

```cpp
update_field([this]() -> auto& { return cache_->value; },
    [this]() { return (*face)["value"_s]; });
```

3. The documentation in the comments have been updated to match the
   changes.