Added double_range namespace
`double_range` is a namepsace that defines classes that implicitly
convert to `double` on all operations but in the constructor they can
have a check to ensure a specific condition on the value is satisfied,
throwing an error when it isn't.

`Unit` is the only such class so far which checks that the value is
between 0 and 1 upon construction.