Test comparison of refactored and old Penman-Monteith classes
A google test is added here to compare the output of the refactored
Penman-Monteith ET calculation with the old version. At this commit the
tests are successful.

There are three test categories:

StomatalResistanceComparison: goes through a couple of sample cases,
testing ET and stomatal resistance

BoundaryConditionsComparison

Tries a range of input variable values.

PerformanceComparison

Checks that the refactored version is at worst 1% slower than the old
one.